requirements.txt
numpy>=1.26.0
pandas>=2.1.0
matplotlib>=3.8.0
scikit-learn>=1.3.0
tensorflow-macos>=2.15.0
pillow>=10.0.0
flask>=2.3.2
pytest>=7.3.1
tqdm>=4.65.0
pyttsx3>=2.90
scipy>=1.11.0
opencv-python>=4.8.0
seaborn>=0.12.2